Fix ClassCastException in download from current playlist
SqueezerCurrentPlaylist can't be cast to SqueezerAbstractSongList,
because it no longer extends SqueezerAbstractPlaylist (as the callback
requires special handling of the currently playing song).
The downloadSong method is moved to SqueezerBaseActivity thus
eliminating the need to cast. (and the need to duplicate it in the
music folder list).
