+++ src/uk/org/ngo/squeezer/itemlists/SqueezerMusicFolderView.java
@@ -117,8 +117,7 @@ public class SqueezerMusicFolderView extends SqueezerBaseItemView<SqueezerMusicF
             throws RemoteException {
         switch (menuItem.getItemId()) {
             case R.id.download:
-                ((SqueezerMusicFolderListActivity) getActivity())
-                        .downloadSong(selectedItem.getId());
+                getActivity().downloadSong(selectedItem.getId());
                 return true;
         }
         return super.doItemContext(menuItem, index, selectedItem);
